+++ microservices/services/file/internal/dao/system/setting_alias.go
@@ -1,0 +1,7 @@
+package system
+
+import sharedsetting "github.com/go-admin-kit/services/shared/pkg/setting"
+
+type SettingDAO = sharedsetting.SettingDAO
+
+var NewSettingDAO = sharedsetting.NewSettingDAO